Fix reading backlight level on 64-bit systems (bug 63205)
obviously incorrect output of "xbacklight -get":
51671552015166906368.000000

valgrind reports an invalid read of size 8 at xbacklight.c:91

https://bugs.freedesktop.org/show_bug.cgi?id=63205

+++ xbacklight.c
@@ -87,7 +87,7 @@ backlight_get (xcb_connection_t *conn, xcb_randr_output_t output)
 	prop_reply->format != 32) {
 	value = -1;
     } else {
-	value = *((long *) xcb_randr_get_output_property_data (prop_reply));
+	value = *((int32_t *) xcb_randr_get_output_property_data (prop_reply));
     }
 
     free (prop_reply);
